Check for APPOINTED TxStatus before order cancel
During redirect checkouts it might happen, that the customer hits the back / error action by clicking browser back.

To prevent cancelation for orders that already have been appointed, this commit adds a check for APPOINTED TxStatus. If the Transaction has been APPOINTED the order will not be canceled and the quote will not be reactivated.