Update Checkout.php
Under some circumstances, there is the error "Sorry, your transaction with Amazon Pay was not successful. Please try again." thrown, even if the values seem to be identical. After implementing a logging, we figured out, that there seems to be a comparison between a string and a floating point number, which is not safe anyway. See: https://stackoverflow.com/questions/3148937/compare-floats-in-php

To avoid this, the value is converted to a string before writing the value into the session and to convert both values to the same "number_format" value with 4 digits